+++ src/App.tsx
@@ -12,6 +12,7 @@ import { PesananDetailItem } from './pages/Pesanan/PesananDetailItem';
 import { ProduksiList } from './pages/Produksi/ProduksiList';
 import { FinanceList } from './pages/Keuangan/FinanceList';
 import { FinanceDetail } from './pages/Keuangan/FinanceDetail';
+import { AktivitasList } from './pages/Aktivitas/AktivitasList';
 
 
 function App() {
@@ -34,6 +35,7 @@ function App() {
           <Route index element={<FinanceList />} />
           <Route path=":id" element={<FinanceDetail />} />
         </Route>
+        <Route path="aktivitas" element={<AktivitasList />} />
 
       </Route>
     </Routes>